refactor(systest): split the systest lib into stage directories
Mechanical restructuring of nes-systest-lib into per-stage directories,
without behavior or type changes: the parser moves to Parser/, the
configuration and the argument parsing (split out of SystestStarter) to
Config/, discovery (split out of SystestState) to Discovery/, the runner
and query submitter to Runner/, the result checking to ResultChecker/,
and the progress tracker and logging setup become Progress and Logging.
main() shrinks to src/Main.cpp. A new Discovery/TestFileReader is now
the only place a test file is opened; the parser loses loadFile and
always receives text via loadString.
