+++ nes-systests/systest/include/Discovery/TestDiscovery.hpp
@@ -1,0 +1,29 @@
+/*
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        https://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+
+#pragma once
+
+#include <ostream>
+#include <Config/Config.hpp>
+#include <SystestState.hpp>
+
+namespace NES::Systest
+{
+
+/// load test file map objects from files defined in systest config
+TestFileMap loadTestFileMap(const SystestConfiguration& config);
+
+std::ostream& operator<<(std::ostream& os, const TestFileMap& testMap);
+
+}
