+++ nes-systests/systest/include/Logging.hpp
@@ -1,0 +1,26 @@
+/*
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        https://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+
+#pragma once
+
+#include <Config/Config.hpp>
+
+namespace NES
+{
+
+/// Sets up the systest log file (a generated name under the build directory unless the configuration names one)
+/// and points the latest.log symlink at it.
+void setupLogging(const SystestConfiguration& config);
+
+}
