+++ nes-systests/systest/src/Discovery/TestFileReader.cpp
@@ -1,0 +1,40 @@
+/*
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        https://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+
+#include <Discovery/TestFileReader.hpp>
+
+#include <filesystem>
+#include <fstream>
+#include <sstream>
+#include <string>
+
+#include <ErrorHandling.hpp>
+
+namespace NES
+{
+
+std::string readTestFile(const std::filesystem::path& path)
+{
+    const std::ifstream file(path);
+    if (not file.is_open() or file.bad())
+    {
+        throw TestException("could not read test file {}", path.string());
+    }
+
+    std::stringstream buffer;
+    buffer << file.rdbuf();
+    return buffer.str();
+}
+
+}
