+++ nes-systests/systest/src/Logging.cpp
@@ -1,0 +1,103 @@
+/*
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        https://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+
+#include <Logging.hpp>
+
+#include <chrono>
+#include <cstdlib>
+#include <filesystem>
+#include <iostream>
+#include <string>
+#include <system_error>
+#include <unistd.h>
+#include <Config/Config.hpp>
+#include <Util/Logger/LogLevel.hpp>
+#include <Util/Logger/impl/NesLogger.hpp>
+#include <fmt/format.h>
+#include <fmt/ostream.h>
+
+namespace NES
+{
+namespace
+{
+void createSymlink(const std::filesystem::path& absoluteLogPath, const std::filesystem::path& symlinkPath)
+{
+    std::error_code errorCode;
+    const auto relativeLogPath = relative(absoluteLogPath, symlinkPath.parent_path(), errorCode);
+    if (errorCode)
+    {
+        std::cerr << "Error calculating relative path during logger setup: " << errorCode.message() << "\n";
+        return;
+    }
+
+    if (exists(symlinkPath, errorCode) || is_symlink(symlinkPath, errorCode))
+    {
+        std::filesystem::remove(symlinkPath, errorCode);
+        if (errorCode)
+        {
+            std::cerr << "Error removing existing symlink during logger setup:  " << errorCode.message() << "\n";
+        }
+    }
+
+    try
+    {
+        create_symlink(relativeLogPath, symlinkPath);
+    }
+    catch (const std::filesystem::filesystem_error& e)
+    {
+        std::cerr << "Error creating symlink during logger setup: " << e.what() << '\n';
+    }
+}
+}
+
+void setupLogging(const SystestConfiguration& config)
+{
+    std::filesystem::path absoluteLogPath;
+    const std::filesystem::path logDir = std::filesystem::path(PATH_TO_BINARY_DIR) / "nes-systests";
+
+    if (config.logFilePath.getValue().empty())
+    {
+        std::error_code errorCode;
+        create_directories(logDir, errorCode);
+        if (errorCode)
+        {
+            std::cerr << "Error creating log directory during logger setup: " << errorCode.message() << "\n";
+            return;
+        }
+
+        const auto now = std::chrono::system_clock::now();
+        const auto pid = ::getpid();
+        const std::string logFileName = fmt::format("SystemTest_{:%Y-%m-%d_%H-%M-%S}_{:d}.log", now, pid);
+
+        absoluteLogPath = logDir / logFileName;
+    }
+    else
+    {
+        absoluteLogPath = config.logFilePath.getValue();
+        const std::filesystem::path parentDir = absoluteLogPath.parent_path();
+        if (not exists(parentDir) or not is_directory(parentDir))
+        {
+            fmt::println(std::cerr, "Error creating log file during logger setup: directory does not exist: file://{}", parentDir.string());
+            std::exit(1); /// NOLINT(concurrency-mt-unsafe)
+        }
+    }
+
+    fmt::println(std::cout, "Find the log at: file://{}", absoluteLogPath.string());
+    Logger::setupLogging(absoluteLogPath.string(), LogLevel::LOG_DEBUG, false);
+
+    const auto symlinkPath = logDir / "latest.log";
+    createSymlink(absoluteLogPath, symlinkPath);
+}
+
+}
