+++ nes-systests/systest/tests/Parser/SystestParserInvalidTestFilesTests.cpp
@@ -17,12 +17,13 @@
 #include <utility>
 #include <vector>
 
+#include <Discovery/TestFileReader.hpp>
+#include <Parser/SystestParser.hpp>
 #include <Util/Logger/Logger.hpp>
 #include <Util/Logger/impl/NesLogger.hpp>
 #include <gtest/gtest.h>
 #include <BaseUnitTest.hpp>
 #include <ErrorHandling.hpp>
-#include <SystestParser.hpp>
 #include <SystestState.hpp>
 
 namespace NES::Systest
@@ -45,7 +46,7 @@ TEST_F(SystestParserInvalidTestFilesTest, InvalidTestFile)
     GTEST_FLAG_SET(death_test_style, "threadsafe");
     const std::string filename = SYSTEST_DATA_DIR "invalid.dummy";
     SystestParser parser{};
-    ASSERT_TRUE(parser.loadFile(filename));
+    ASSERT_TRUE(parser.loadString(NES::readTestFile(filename)));
     ASSERT_EXCEPTION_ERRORCODE({ parser.parse(); }, ErrorCode::SLTUnexpectedToken)
 }
 
@@ -64,7 +65,7 @@ TEST_F(SystestParserInvalidTestFilesTest, InvalidErrorCodeTest)
             /// nop, ensure parsing
         });
 
-    ASSERT_TRUE(parser.loadFile(filename));
+    ASSERT_TRUE(parser.loadString(NES::readTestFile(filename)));
     ASSERT_EXCEPTION_ERRORCODE({ parser.parse(); }, ErrorCode::SLTUnexpectedToken)
 }
 
@@ -83,7 +84,7 @@ TEST_F(SystestParserInvalidTestFilesTest, InvalidErrorMessageTest)
             /// nop, ensure parsing
         });
 
-    ASSERT_TRUE(parser.loadFile(filename));
+    ASSERT_TRUE(parser.loadString(NES::readTestFile(filename)));
     ASSERT_EXCEPTION_ERRORCODE({ parser.parse(); }, ErrorCode::SLTUnexpectedToken)
 }
 
@@ -96,7 +97,7 @@ TEST_F(SystestParserInvalidTestFilesTest, InvalidTokenTest)
     parser.registerOnCreateCallback(
         [&](const std::string&, const std::optional<std::pair<TestDataIngestionType, std::vector<std::string>>>&) { });
 
-    ASSERT_TRUE(parser.loadFile(filename));
+    ASSERT_TRUE(parser.loadString(NES::readTestFile(filename)));
     ASSERT_EXCEPTION_ERRORCODE({ parser.parse(); }, ErrorCode::SLTUnexpectedToken)
 }
 
@@ -112,7 +113,7 @@ TEST_F(SystestParserInvalidTestFilesTest, InvalidDifferentialTest)
     parser.registerOnDifferentialQueryBlockCallback(
         [](std::string, std::string, SystestQueryId, SystestQueryId) { /* nop, ensure parsing*/ });
 
-    ASSERT_TRUE(parser.loadFile(filename));
+    ASSERT_TRUE(parser.loadString(NES::readTestFile(filename)));
     ASSERT_EXCEPTION_ERRORCODE({ parser.parse(); }, ErrorCode::SLTUnexpectedToken)
 }
